0.6.0
* Implement HLBVH for scene objects and lights
* Implement environment maps for scene
* Make kd-tree traversal iterative
* Remove material from kd-tree, make triangles store reference to their material
* Allow instancing objects without cloning underlying object
* Implement RGB to spectrum conversion due to Jacob & Hanika
* Remove infinite planes
* Add Amazon lumberyard bistro example
* Add white balancing to film
* Make mediums scale with the scene
* Make uvs repeat when outside range [0,1]
* Weigh visible spectrum in wavelenght sampling
* Support indexed and grayscale png decoding
* Add HDRI image decoding
* Improve memory footprint of parser
* Support more arguments in material parser
* Add support for normal maps

+++ src/math/complex.rs
@@ -7,14 +7,12 @@ use std::ops::{
 mod complex_tests;
 
 #[derive(Clone, Copy)]
-#[allow(non_snake_case)]
 pub struct Complex {
     Re: Float,
     Im: Float,
 }
 
 impl Complex {
-    #[allow(non_snake_case)]
     #[inline]
     pub fn new(Re: Float, Im: Float) -> Self {
         Self { Re, Im }
@@ -51,7 +49,7 @@ impl Complex {
 
     #[inline]
     fn arg(&self) -> Float {
-        libm::atan2(self.Im, self.Re)
+        libm::atan2(self.Im as f64, self.Re as f64) as Float
     }
 }
